dbr quant overhead[5/x]: remove unnecessary asserts (#68370)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68370

Removes asserts which are duplicate (the same condition is checked
when calculating the hook type, so there is no need to check it again).
For the assert in `validate_is_at_last_seen_idx`, rewrites it to
raise an Error instead to ensure it does not get stripped in
production environments.

Test Plan:
```
python test/test_quantization.py TestQuantizeDBR
```

Reviewed By: jerryzh168

Differential Revision: D32463766

Pulled By: vkuzo

fbshipit-source-id: 8a7b7e0bf270bc327f49bd3e5bd156339e846381