CMake: Update precompiled header and fix support (#67851)
Summary:
This fixes the `USE_PRECOMPILED_HEADERS` cmake version check which was accidentally inverted, so it was always disabled.

I've also made the precompiled header so it only includes headers used in 95% or more of code, weighted by compile time. This limits it to the standard library, `c10` and a limited subset of `ATen/core`. Crucially, the new pch doesn't depend on `native_functions.yaml` so won't cause as much unnecessary rebuilding.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/67851

Reviewed By: zou3519

Differential Revision: D32290902

Pulled By: dagitses

fbshipit-source-id: dfc33330028c99b02ff40963926c1f1260d00d00