[quant][core][gpu] Added additional uid-data_ptr pair for broadcasted_bias in quantized cudnn linear op
Summary:
An additional entry for broadcasted_bias in the uid-data_ptr pairing
appears to be needed in the current version of cudnn. Without it, some
test cases are failing. NVIDIA is currently investigating this issue.
When this issue is fixd, we can remove the changes in this commit.

Test plan:
```
python test/test_quantization.py -k test_qlinear_cudnn
```

Pull Request resolved: https://github.com/pytorch/pytorch/pull/75969

Approved by: https://github.com/jerryzh168